Expressions to Swift, cont.
OK, a lot more is working in the unit tests, like 90% or more. I think the rest is mostly just cleaning up expected results, since some of those have changed. I also improved the precedence operators, so those are working better in that expression are parsing better with equalities and operators intermixed.