More work on Swift expression
OK, things are looking even better. I've slow "deprecating" AJRExpression. As such, we now introduce the idea of a literal that's resolved via the context. This is the key feature I need for variables to work. A bunch of this was done by making things implement AJREvaluation rather than subclassing AJRExpression, which is why I'm eliminating AJRExpression. So, in the long run, ideally we won't have an "Expression" superclass, we'll just need the subclasses, which won't rely on the superclass. I'll also probably rename AJRSimpleExpression to AJRCompoundExpression, but keep AJRUnaryExpression and AJRFunctionExpression as is.