Fixed an issue with the root object / context evaluation of key paths
Basically, if our context had a root object, when encountering a key path, we needed to try and evaluation on the root object. Otherwise, we needed to evaluate on the context. This is somewhat of a two ways to do the same thing, but contexts can be more complicated, so think of them as the advanced case.