Removed an unnecessary check
Basically, if you're registering an extension point, you only have to define the class name if you define a registry selector, but I was making it so a missing class name caused the extension point to not get registered. Note we still have to verify against name, since otherwise we can't register the extension point.